requirements.txt
# Core
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
openai==1.12.0
python-dotenv==1.0.0

# Graph
networkx==3.2.1
pyvis==0.3.2
matplotlib==3.8.2

# Vector Store
chromadb==0.4.22

# UI
streamlit==1.31.0
plotly==5.18.0

# Utilities
pydantic==2.5.0
numpy==1.26.3
pandas==2.1.4
